pyproject.toml
[tool.poetry]
name = "alekorn-gendiff"
version = "0.3.3"
description = ""
authors = ["alekorn <entkorn@gmail.com>"]
packages = [
  { include = "gendiff" },
]
[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3"
pylint = "^2.4.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.4"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

